requirements.txt
fastapi
uvicorn[standard]
pydantic
pyyaml
langchain-core
langchain-chroma
langchain-google-genai
chromadb
python-dotenv
